remove mocks from build
